v19/cloudconfig: extract baseExtension type to its own file
Follow up of https://github.com/giantswarm/aws-operator/pull/1200#discussion_r229022868
